Fix line parsing in cases like "FREQ=DAILY"
Google calendar with a daily rrule event wouldn't be found in a query.
GetSubProperty searched until a ;, which didn't exist, so find_first_of returned npos (-1), leading to errors. The line was consumed fully, including the CR, so ConvertFrequency(string) couldn't properly match the freq string to a freq enum